requirements.txt
pytz
requests
redis>=3.3.6
pycryptodome==3.12.0
xmltodict==0.12.0